feat: Add gradient support to FeatureCard
This commit introduces a gradient background option for the `FeatureCard` widget.

The `AppFeatures` enum and `FeatureCard` widget have been updated to include an optional `gradient` parameter. When provided, this gradient will be applied to the background of the card.

The "Transcription" feature card has been updated to utilize this new functionality, now displaying a green to blue linear gradient.